+++ packages/gasket-cli/src/hooks/default-plugins.js
@@ -1,4 +1,5 @@
 module.exports = [
   require('@gasket/command-plugin'),
-  require('@gasket/lifecycle-plugin')
+  require('@gasket/lifecycle-plugin'),
+  require('@gasket/metadata-plugin')
 ];