+++ packages/gasket-git-plugin/test/prompt.test.js
@@ -25,6 +25,7 @@ describe('prompt', function () {
     mockContext.gitInit = false;
     await prompt({}, mockContext, mockUtils);
 
+    // eslint-disable-next-line require-atomic-updates
     mockContext.gitInit = true;
     await prompt({}, mockContext, mockUtils);